backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.31
psycopg2-binary==2.9.11
python-dotenv==1.2.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
alembic==1.16.4
slowapi==0.1.9
psutil==7.0.0
httpx
paramiko==3.5.0
websockets==13.1
apscheduler==3.10.4
